refactor: move instant search wrapper
