+++ libs/journeys/ui/src/libs/algolia/InstantSearchWrapper.tsx
